create some api and code base for create signalr

+++ HandHubAPI/Domain/Entities/PriceNegotiationEntity.cs
@@ -9,14 +9,14 @@ public class PriceNegotiationEntity : BaseEntity
 
     public int BuyerId { get; set; }
 
-    public int OfferPrice { get; set; }
+    public decimal OfferPrice { get; set; }
 
-    public int Status { get; set; }
+    public int Status { get; set; } // 0: Wait, 1: Accepted, 2: Rejected
 
     [MaxLength(50)]
     public string SellerResponse { get; set; } = string.Empty;
 
-    public int FinalPrice { get; set; }
+    public decimal FinalPrice { get; set; }
     [ForeignKey("ProductId")]
     public ProductEntity? Product { get; set; }
     [ForeignKey("BuyerId")]
